debug

+++ label_distance.py
@@ -98,7 +98,7 @@ def indication_distance(args):
     label_type, count = np.unique(labels_oh_str, axis=0, return_counts=True)
     count_sort_idx = np.argsort(-count)
     label_type = label_type[count_sort_idx]
-    target_fair_labels = label_type[:1].astype(int)
+    target_fair_labels = label_type[:1]
     print(target_fair_labels, count[count_sort_idx])
 
     labels_express = {}
